Now excerpts works as they should
