+++ lib/premonition.rb
@@ -1,5 +1,6 @@
 require 'jekyll'
 require 'premonition/version'
+require 'premonition/processor'
 require 'premonition/resources'
 require 'premonition/hook'
